Do not remove the `/usr/local/elasticsearch` directory when it's a symlink.
This means it was created by Ark, not by Chef when creating the user.

h/t: @avishai-ish-shalom

Related: #104

+++ recipes/default.rb
@@ -26,6 +26,8 @@ end
 bash "remove the elasticsearch user home" do
   user    'root'
   code    "rm -rf  #{node.elasticsearch[:dir]}/elasticsearch"
+
+  not_if  "test -L #{node.elasticsearch[:dir]}/elasticsearch"
   only_if "test -d #{node.elasticsearch[:dir]}/elasticsearch"
 end
